+++ point-of-sale-sdk/src/test/java/com/squareup/sdk/pos/RealPosClientTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.squareup.sdk.register;
+package com.squareup.sdk.pos;
 
 import android.content.ActivityNotFoundException;
 import android.content.Context;
@@ -36,23 +36,21 @@ import org.robolectric.RobolectricTestRunner;
 
 import static android.content.Intent.ACTION_VIEW;
 import static android.content.Intent.FLAG_ACTIVITY_NEW_TASK;
-import static com.squareup.sdk.register.ChargeRequest.TenderType.CARD;
-import static com.squareup.sdk.register.CurrencyCode.USD;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_API_VERSION;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_AUTO_RETURN_TIMEOUT_MS;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_CURRENCY_CODE;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_CUSTOMER_ID;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_LOCATION_ID;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_NOTE;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_REGISTER_CLIENT_ID;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_REQUEST_METADATA;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_SDK_VERSION;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_TENDER_CARD;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_TENDER_TYPES;
-import static com.squareup.sdk.register.RegisterApi.EXTRA_TOTAL_AMOUNT;
-import static com.squareup.sdk.register.RegisterApi.INTENT_ACTION_CHARGE;
-import static com.squareup.sdk.register.TestData.INVALID_SIGNATURE;
-import static com.squareup.sdk.register.TestData.REGISTER_SIGNATURE;
+import static com.squareup.sdk.pos.PosApi.EXTRA_API_VERSION;
+import static com.squareup.sdk.pos.PosApi.EXTRA_AUTO_RETURN_TIMEOUT_MS;
+import static com.squareup.sdk.pos.PosApi.EXTRA_CURRENCY_CODE;
+import static com.squareup.sdk.pos.PosApi.EXTRA_CUSTOMER_ID;
+import static com.squareup.sdk.pos.PosApi.EXTRA_LOCATION_ID;
+import static com.squareup.sdk.pos.PosApi.EXTRA_NOTE;
+import static com.squareup.sdk.pos.PosApi.EXTRA_POINT_OF_SALE_CLIENT_ID;
+import static com.squareup.sdk.pos.PosApi.EXTRA_REQUEST_METADATA;
+import static com.squareup.sdk.pos.PosApi.EXTRA_SDK_VERSION;
+import static com.squareup.sdk.pos.PosApi.EXTRA_TENDER_CARD;
+import static com.squareup.sdk.pos.PosApi.EXTRA_TENDER_TYPES;
+import static com.squareup.sdk.pos.PosApi.EXTRA_TOTAL_AMOUNT;
+import static com.squareup.sdk.pos.PosApi.INTENT_ACTION_CHARGE;
+import static com.squareup.sdk.pos.TestData.INVALID_SIGNATURE;
+import static com.squareup.sdk.pos.TestData.POINT_OF_SALE_SIGNATURE;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Matchers.anyInt;
@@ -63,28 +61,28 @@ import static org.mockito.MockitoAnnotations.initMocks;
 
 @SuppressWarnings({ "WrongConstant", "deprecation" }) //
 @RunWith(RobolectricTestRunner.class) //
-public class RealRegisterClientTest {
+public class RealPosClientTest {
 
   private static final String CLIENT_ID = "CLIENT_ID";
 
   @Mock(answer = Answers.RETURNS_DEEP_STUBS) Context context;
   List<ResolveInfo> chargeActivities;
-  RealRegisterClient client;
+  RealPosClient client;
 
   @Before public void setUp() {
     initMocks(this);
     chargeActivities = new ArrayList<>();
 
-    installApp("com.squareup", 2, REGISTER_SIGNATURE);
+    installApp("com.squareup", 2, POINT_OF_SALE_SIGNATURE);
 
     when(context.getPackageManager()
         .queryIntentActivities(new Intent(INTENT_ACTION_CHARGE), 0)).thenReturn(chargeActivities);
 
-    client = new RealRegisterClient(context, CLIENT_ID);
+    client = new RealPosClient(context, CLIENT_ID);
   }
 
   @Test public void opensPlayStoreWhenInstalled() throws Exception {
-    client.openRegisterPlayStoreListing();
+    client.openPointOfSalePlayStoreListing();
 
     Intent expectedIntent = new Intent(ACTION_VIEW, Uri.parse("market://details?id=com.squareup"));
     expectedIntent.addFlags(FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_WHEN_TASK_RESET);
@@ -95,7 +93,7 @@ public class RealRegisterClientTest {
     when(context.getPackageManager().getPackageInfo("com.android.vending", 0)).thenThrow(
         new PackageManager.NameNotFoundException());
 
-    client.openRegisterPlayStoreListing();
+    client.openPointOfSalePlayStoreListing();
 
     Intent expectedIntent = new Intent(ACTION_VIEW,
         Uri.parse("https://play.google.com/store/apps/details?id=com.squareup"));
@@ -103,54 +101,55 @@ public class RealRegisterClientTest {
     verify(context).startActivity(expectedIntent);
   }
 
-  @Test public void launchRegister() {
-    client.launchRegister();
+  @Test public void launchPointOfSale() {
+    client.launchPointOfSale();
 
     Intent expectedIntent = context.getPackageManager().getLaunchIntentForPackage("com.squareup");
     verify(context).startActivity(expectedIntent);
   }
 
-  @Test public void registerInstalled() {
-    assertThat(client.isRegisterInstalled()).isTrue();
+  @Test public void pointOfSaleInstalled() {
+    assertThat(client.isPointOfSaleInstalled()).isTrue();
   }
 
-  @Test public void registerNotInstalled() {
+  @Test public void pointOfSaleNotInstalled() {
     chargeActivities.clear();
 
-    assertThat(client.isRegisterInstalled()).isFalse();
+    assertThat(client.isPointOfSaleInstalled()).isFalse();
   }
 
   @Test public void betaFlavorInstalled() {
     chargeActivities.clear();
-    installApp("com.squareup.beta", 2, REGISTER_SIGNATURE);
+    installApp("com.squareup.beta", 2, POINT_OF_SALE_SIGNATURE);
 
-    assertThat(client.isRegisterInstalled()).isTrue();
+    assertThat(client.isPointOfSaleInstalled()).isTrue();
   }
 
   @Test public void invalidPackagePrefix() {
     chargeActivities.clear();
-    installApp("com.triangleup", 2, REGISTER_SIGNATURE);
+    installApp("com.triangleup", 2, POINT_OF_SALE_SIGNATURE);
 
-    assertThat(client.isRegisterInstalled()).isFalse();
+    assertThat(client.isPointOfSaleInstalled()).isFalse();
   }
 
   @Test public void invalidSignature() {
     chargeActivities.clear();
     installApp("com.squareup", 2, INVALID_SIGNATURE);
 
-    assertThat(client.isRegisterInstalled()).isFalse();
+    assertThat(client.isPointOfSaleInstalled()).isFalse();
   }
 
   @Test public void multipleFlavorsInstalled() {
     chargeActivities.clear();
-    installApp("com.squareup", 2, REGISTER_SIGNATURE);
-    installApp("com.squareup.beta", 3, REGISTER_SIGNATURE);
+    installApp("com.squareup", 2, POINT_OF_SALE_SIGNATURE);
+    installApp("com.squareup.beta", 3, POINT_OF_SALE_SIGNATURE);
 
-    assertThat(client.isRegisterInstalled()).isTrue();
+    assertThat(client.isPointOfSaleInstalled()).isTrue();
   }
 
   @Test public void intentCreatedWithRequestParams() {
-    ChargeRequest request = new ChargeRequest.Builder(1_00, USD).restrictTendersTo(CARD)
+    ChargeRequest request = new ChargeRequest.Builder(1_00, CurrencyCode.USD).restrictTendersTo(
+        ChargeRequest.TenderType.CARD)
         .autoReturn(4, SECONDS)
         .enforceBusinessLocation("location")
         .customerId("customerId")
@@ -162,9 +161,9 @@ public class RealRegisterClientTest {
 
     assertThat(intent.getIntExtra(EXTRA_TOTAL_AMOUNT, -1)).isEqualTo(1_00);
     assertThat(intent.getStringExtra(EXTRA_CURRENCY_CODE)).isEqualTo("USD");
-    assertThat(intent.getStringExtra(EXTRA_REGISTER_CLIENT_ID)).isEqualTo(CLIENT_ID);
+    assertThat(intent.getStringExtra(EXTRA_POINT_OF_SALE_CLIENT_ID)).isEqualTo(CLIENT_ID);
     assertThat(intent.getStringExtra(EXTRA_NOTE)).isEqualTo("note");
-    assertThat(intent.getStringExtra(EXTRA_API_VERSION)).isEqualTo("v1.2");
+    assertThat(intent.getStringExtra(EXTRA_API_VERSION)).isEqualTo("v2.0");
     assertThat(intent.getStringExtra(EXTRA_REQUEST_METADATA)).isEqualTo("metadata");
     assertThat(intent.getStringExtra(EXTRA_LOCATION_ID)).isEqualTo("location");
     assertThat(intent.getStringExtra(EXTRA_CUSTOMER_ID)).isEqualTo("customerId");
@@ -175,11 +174,11 @@ public class RealRegisterClientTest {
   }
 
   @Test public void fieldsSetAsExpectedForDefaultRequest() {
-    ChargeRequest request = new ChargeRequest.Builder(1_00, USD).build();
+    ChargeRequest request = new ChargeRequest.Builder(1_00, CurrencyCode.USD).build();
 
     Intent intent = client.createChargeIntent(request);
 
-    assertThat(intent.hasExtra(EXTRA_REGISTER_CLIENT_ID)).isTrue();
+    assertThat(intent.hasExtra(EXTRA_POINT_OF_SALE_CLIENT_ID)).isTrue();
     assertThat(intent.hasExtra(EXTRA_TOTAL_AMOUNT)).isTrue();
     assertThat(intent.hasExtra(EXTRA_API_VERSION)).isTrue();
     assertThat(intent.hasExtra(EXTRA_SDK_VERSION)).isTrue();
@@ -194,37 +193,39 @@ public class RealRegisterClientTest {
 
   @Test public void pinsToHighestVersionNumber() {
     chargeActivities.clear();
-    installApp("com.squareup", 2, REGISTER_SIGNATURE);
-    installApp("com.squareup.beta", 3, REGISTER_SIGNATURE);
+    installApp("com.squareup", 2, POINT_OF_SALE_SIGNATURE);
+    installApp("com.squareup.beta", 3, POINT_OF_SALE_SIGNATURE);
 
-    Intent intent = client.createChargeIntent(new ChargeRequest.Builder(1_00, USD).build());
+    Intent intent =
+        client.createChargeIntent(new ChargeRequest.Builder(1_00, CurrencyCode.USD).build());
 
     assertThat(intent.getPackage()).isEqualTo("com.squareup.beta");
   }
 
-  @Test(expected = ActivityNotFoundException.class) public void createIntentWithNoRegisterThrows() {
+  @Test(expected = ActivityNotFoundException.class)
+  public void createIntentWithNoPointOfSaleThrows() {
     chargeActivities.clear();
-    client.createChargeIntent(new ChargeRequest.Builder(1_00, USD).build());
+    client.createChargeIntent(new ChargeRequest.Builder(1_00, CurrencyCode.USD).build());
   }
 
   @Test(expected = ActivityNotFoundException.class)
-  public void createIntentWithInvalidRegisterThrows() {
+  public void createIntentWithInvalidPointOfSaleThrows() {
     chargeActivities.clear();
     installApp("com.squareup", 2, INVALID_SIGNATURE);
-    client.createChargeIntent(new ChargeRequest.Builder(1_00, USD).build());
+    client.createChargeIntent(new ChargeRequest.Builder(1_00, CurrencyCode.USD).build());
   }
 
   @Test(expected = ActivityNotFoundException.class)
-  public void launchRegisterThrowsWithNoRegisterInstalled() {
+  public void launchPointOfSaleThrowsWithNoPointOfSaleInstalled() {
     chargeActivities.clear();
-    client.launchRegister();
+    client.launchPointOfSale();
   }
 
   @Test(expected = ActivityNotFoundException.class)
-  public void launchRegisterThrowsWithInvalidSignature() {
+  public void launchPointOfSaleThrowsWithInvalidSignature() {
     chargeActivities.clear();
     installApp("com.squareup", 2, INVALID_SIGNATURE);
-    client.launchRegister();
+    client.launchPointOfSale();
   }
 
   private void installApp(String packageName, int versionCode, Signature signature) {