+++ sample-bikeshop/src/main/java/com/example/owensbikes/TransactionSuccessActivity.java
@@ -21,9 +21,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
-import android.widget.Button;
 import android.widget.TextView;
-import com.squareup.sdk.register.ChargeRequest;
 
 /**
  * Activity for displaying the result of a successful transaction. It receives the order number